@@ -1,16 +1,16 @@
 name: jverify
 description: JIGUANG Official Jverifycation SDK flutter plugin project.
-version: 0.6.12
+version: 2.1.2
 homepage: https://www.jiguang.cn
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=1.20.0"
+
 
 dependencies:
-  platform: ^2.0.0
   flutter:
     sdk: flutter
-
 
 
 dev_dependencies:
@@ -26,12 +26,16 @@ flutter:
   # The androidPackage and pluginClass identifiers should not ordinarily
   # be modified. They are used by the tooling to maintain consistency when
   # adding or updating assets for this project.
+#  plugin:
+#    androidPackage: com.jiguang.jverify
+#    pluginClass: JverifyPlugin
   plugin:
-    android:
-      package: com.jiguang.jverify
-      pluginClass: JverifyPlugin
-    ios:
-      pluginClass: JverifyPlugin
+    platforms:
+      android:
+        package: com.jiguang.jverify
+        pluginClass: JverifyPlugin
+      ios:
+        pluginClass: JverifyPlugin
 
   # To add assets to your plugin package, add an assets section, like this:
   # assets:
