@@ -1,21 +1,16 @@
 import 'dart:async';
-
 import 'dart:io';
+
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-import 'package:modal_progress_hud/modal_progress_hud.dart';
 import 'package:jverify/jverify.dart';
 
+import 'load.dart';
 
-
-
-void main() => runApp(
-    new MaterialApp(
+void main() => runApp(new MaterialApp(
       title: "demo",
       theme: new ThemeData(primaryColor: Colors.white),
       home: MyApp(),
-    )
-);
+    ));
 
 class MyApp extends StatefulWidget {
   @override
@@ -23,24 +18,22 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-
   /// 统一 key
   final String f_result_key = "result";
+
   /// 错误码
-  final  String  f_code_key = "code";
+  final String f_code_key = "code";
+
   /// 回调的提示信息，统一返回 flutter 为 message
-  final  String  f_msg_key  = "message";
+  final String f_msg_key = "message";
+
   /// 运营商信息
-  final  String  f_opr_key  = "operator";
+  final String f_opr_key = "operator";
 
-
-  String _platformVersion = 'Unknown';
   String _result = "token=";
   var controllerPHone = new TextEditingController();
   final Jverify jverify = new Jverify();
-  bool _loading = false;
-  String _token;
-
+  String? _token;
 
   @override
   void initState() {
@@ -55,7 +48,7 @@ class _MyAppState extends State<MyApp> {
         appBar: AppBar(
           title: const Text('JVerify example'),
         ),
-        body: ModalProgressHUD(child: _buildContent(), inAsyncCall: _loading),
+        body: _buildContent(),
       ),
     );
   }
@@ -78,14 +71,13 @@ class _MyAppState extends State<MyApp> {
               mainAxisAlignment: MainAxisAlignment.center,
               children: <Widget>[
                 new CustomButton(
-                    onPressed: (){
+                    onPressed: () {
                       isInitSuccess();
                     },
-                    title: "初始化状态"
-                ),
+                    title: "初始化状态"),
                 new Text("   "),
                 new CustomButton(
-                  onPressed: (){
+                  onPressed: () {
                     checkVerifyEnable();
                   },
                   title: "网络环境是否支持",
@@ -164,7 +156,7 @@ class _MyAppState extends State<MyApp> {
       setState(() {
         if (result) {
           _result = "sdk 初始换成功";
-        }else {
+        } else {
           _result = "sdk 初始换失败";
         }
       });
@@ -178,18 +170,17 @@ class _MyAppState extends State<MyApp> {
       setState(() {
         if (result) {
           _result = "当前网络环境【支持认证】！";
-        }else {
+        } else {
           _result = "当前网络环境【不支持认证】！";
         }
       });
     });
   }
 
-
   /// 获取号码认证token
   void getToken() {
     setState(() {
-      _loading = true;
+      _showLoading(context);
     });
     jverify.checkVerifyEnable().then((map) {
       bool result = map[f_result_key];
@@ -199,27 +190,28 @@ class _MyAppState extends State<MyApp> {
           _token = map[f_msg_key];
           String operator = map[f_opr_key];
           setState(() {
-            _loading = false;
+            _hideLoading();
             _result = "[$code] message = $_token, operator = $operator";
           });
         });
       } else {
         setState(() {
-          _loading = false;
+          _hideLoading();
           _result = "[2016],msg = 当前网络环境不支持认证";
         });
       }
     });
   }
+
   /// 获取短信验证码
-  void getSMSCode(){
+  void getSMSCode() {
     setState(() {
-      _loading = true;
+      _showLoading(context);
     });
     String phoneNum = controllerPHone.text;
-    if(phoneNum == null||phoneNum.isEmpty){
+    if (phoneNum == null || phoneNum.isEmpty) {
       setState(() {
-        _loading = false;
+        _hideLoading();
         _result = "[3002],msg = 没有输入手机号码";
       });
       return;
@@ -232,13 +224,13 @@ class _MyAppState extends State<MyApp> {
           int code = map[f_code_key];
           String message = map[f_msg_key];
           setState(() {
-            _loading = false;
+            _hideLoading();
             _result = "[$code] message = $message";
           });
         });
-      }else {
+      } else {
         setState(() {
-          _loading = false;
+          _hideLoading();
           _result = "[3004],msg = 获取短信验证码异常";
         });
       }
@@ -246,9 +238,9 @@ class _MyAppState extends State<MyApp> {
   }
 
   /// 登录预取号
-  void preLogin(){
+  void preLogin() {
     setState(() {
-      _loading = true;
+      _showLoading(context);
     });
     jverify.checkVerifyEnable().then((map) {
       bool result = map[f_result_key];
@@ -258,28 +250,35 @@ class _MyAppState extends State<MyApp> {
           int code = map[f_code_key];
           String message = map[f_msg_key];
           setState(() {
-            _loading = false;
+            _hideLoading();
             _result = "[$code] message = $message";
           });
         });
-      }else {
+      } else {
         setState(() {
-          _loading = false;
+          _hideLoading();
           _result = "[2016],msg = 当前网络环境不支持认证";
         });
       }
     });
   }
 
+  void _showLoading(BuildContext context) {
+    LoadingDialog.show(context);
+  }
+
+  void _hideLoading() {
+    LoadingDialog.hidden();
+  }
+
   /// SDK 请求授权一键登录
   void loginAuth() {
     setState(() {
-      _loading = true;
+      _showLoading(context);
     });
     jverify.checkVerifyEnable().then((map) {
       bool result = map[f_result_key];
       if (result) {
-
         final screenSize = MediaQuery.of(context).size;
         final screenWidth = screenSize.width;
         final screenHeight = screenSize.height;
@@ -290,13 +289,13 @@ class _MyAppState extends State<MyApp> {
         /// ios项目存放在 Assets.xcassets。
         ///
         JVUIConfig uiConfig = JVUIConfig();
-        //uiConfig.authBackgroundImage = ;
+        // uiConfig.authBGGifPath = "main_gif";
 
         //uiConfig.navHidden = true;
         uiConfig.navColor = Colors.red.value;
         uiConfig.navText = "登录";
         uiConfig.navTextColor = Colors.blue.value;
-        uiConfig.navReturnImgPath = "return_bg";//图片必须存在
+        uiConfig.navReturnImgPath = "return_bg"; //图片必须存在
 
         uiConfig.logoWidth = 100;
         uiConfig.logoHeight = 80;
@@ -307,12 +306,14 @@ class _MyAppState extends State<MyApp> {
         uiConfig.logoImgPath = "logo";
 
         uiConfig.numberFieldWidth = 200;
-        uiConfig.numberFieldHeight = 40 ;
+        uiConfig.numberFieldHeight = 40;
         //uiConfig.numFieldOffsetX = isiOS ? 0 : null;//(screenWidth/2 - uiConfig.numberFieldWidth/2).toInt();
         uiConfig.numFieldOffsetY = isiOS ? 20 : 120;
         uiConfig.numberVerticalLayoutItem = JVIOSLayoutItem.ItemLogo;
         uiConfig.numberColor = Colors.blue.value;
         uiConfig.numberSize = 18;
+
+
 
         uiConfig.sloganOffsetY = isiOS ? 20 : 160;
         uiConfig.sloganVerticalLayoutItem = JVIOSLayoutItem.ItemNumber;
@@ -329,21 +330,22 @@ class _MyAppState extends State<MyApp> {
         uiConfig.logBtnText = "登录按钮";
         uiConfig.logBtnTextColor = Colors.brown.value;
         uiConfig.logBtnTextSize = 16;
-        uiConfig.loginBtnNormalImage = "login_btn_normal";//图片必须存在
-        uiConfig.loginBtnPressedImage = "login_btn_press";//图片必须存在
-        uiConfig.loginBtnUnableImage = "login_btn_unable";//图片必须存在
+        uiConfig.loginBtnNormalImage = "login_btn_normal"; //图片必须存在
+        uiConfig.loginBtnPressedImage = "login_btn_press"; //图片必须存在
+        uiConfig.loginBtnUnableImage = "login_btn_unable"; //图片必须存在
 
-        uiConfig.privacyHintToast = true;//only android 设置隐私条款不选中时点击登录按钮默认显示toast。
+        uiConfig.privacyHintToast =
+            true; //only android 设置隐私条款不选中时点击登录按钮默认显示toast。
 
-        uiConfig.privacyState = true;//设置默认勾选
+        uiConfig.privacyState = true; //设置默认勾选
         uiConfig.privacyCheckboxSize = 20;
-        uiConfig.checkedImgPath = "check_image";//图片必须存在
-        uiConfig.uncheckedImgPath = "uncheck_image";//图片必须存在
+        uiConfig.checkedImgPath = "check_image"; //图片必须存在
+        uiConfig.uncheckedImgPath = "uncheck_image"; //图片必须存在
         uiConfig.privacyCheckboxInCenter = true;
         //uiConfig.privacyCheckboxHidden = false;
 
         //uiConfig.privacyOffsetX = isiOS ? (20 + uiConfig.privacyCheckboxSize) : null;
-        uiConfig.privacyOffsetY = 15;// 距离底部距离
+        uiConfig.privacyOffsetY = 15; // 距离底部距离
         uiConfig.privacyVerticalLayoutItem = JVIOSLayoutItem.ItemSuper;
         uiConfig.clauseName = "协议1";
         uiConfig.clauseUrl = "http://www.baidu.com";
@@ -351,12 +353,14 @@ class _MyAppState extends State<MyApp> {
         uiConfig.clauseNameTwo = "协议二";
         uiConfig.clauseUrlTwo = "http://www.hao123.com";
         uiConfig.clauseColor = Colors.red.value;
-        uiConfig.privacyText = ["1极","2光","3认","4证"];
+        uiConfig.privacyText = ["1极", "2光", "3认", "4证"];
         uiConfig.privacyTextSize = 13;
         //uiConfig.privacyWithBookTitleMark = true;
         //uiConfig.privacyTextCenterGravity = false;
-        uiConfig.authStatusBarStyle =  JVIOSBarStyle.StatusBarStyleDarkContent;
+        uiConfig.authStatusBarStyle = JVIOSBarStyle.StatusBarStyleDarkContent;
         uiConfig.privacyStatusBarStyle = JVIOSBarStyle.StatusBarStyleDefault;
+        uiConfig.modelTransitionStyle =
+            JVIOSUIModalTransitionStyle.CrossDissolve;
 
         uiConfig.statusBarColorWithNav = true;
         uiConfig.virtualButtonTransparent = true;
@@ -366,37 +370,52 @@ class _MyAppState extends State<MyApp> {
 
         uiConfig.needStartAnim = true;
         uiConfig.needCloseAnim = true;
+        uiConfig.enterAnim = "activity_slide_enter_bottom";
+        uiConfig.exitAnim = "activity_slide_exit_bottom";
 
-        uiConfig.privacyNavColor =  Colors.red.value;;
+
+        uiConfig.privacyNavColor = Colors.red.value;
         uiConfig.privacyNavTitleTextColor = Colors.blue.value;
         uiConfig.privacyNavTitleTextSize = 16;
 
-        uiConfig.privacyNavTitleTitle  ="ios lai le";//only ios
+        uiConfig.privacyNavTitleTitle = "ios lai le"; //only ios
         uiConfig.privacyNavTitleTitle1 = "协议11 web页标题";
         uiConfig.privacyNavTitleTitle2 = "协议22 web页标题";
-        uiConfig.privacyNavReturnBtnImage = "return_bg";//图片必须存在;
+        uiConfig.privacyNavReturnBtnImage = "return_bg"; //图片必须存在;
+
+        //弹框模式
+        // JVPopViewConfig popViewConfig = JVPopViewConfig();
+        // popViewConfig.width = (screenWidth - 100.0).toInt();
+        // popViewConfig.height = (screenHeight - 150.0).toInt();
+        //
+        // uiConfig.popViewConfig = popViewConfig;
 
         /// 添加自定义的 控件 到授权界面
         List<JVCustomWidget> widgetList = [];
+
         /// 步骤 1：调用接口设置 UI
-        jverify.setCustomAuthorizationView(true, uiConfig, landscapeConfig: uiConfig,widgets: widgetList);
+        jverify.setCustomAuthorizationView(true, uiConfig,
+            landscapeConfig: uiConfig, widgets: widgetList);
 
         /// 步骤 2：调用一键登录接口
 
         /// 方式一：使用同步接口 （如果想使用异步接口，则忽略此步骤，看方式二）
         /// 先，添加 loginAuthSyncApi 接口回调的监听
-        jverify.addLoginAuthCallBackListener((event){
+        jverify.addLoginAuthCallBackListener((event) {
           setState(() {
-            _loading = false;
+            _hideLoading();
+            _hideLoading();
             _result = "监听获取返回数据：[${event.code}] message = ${event.message}";
           });
-          print("通过添加监听，获取到 loginAuthSyncApi 接口返回数据，code=${event.code},message = ${event.message},operator = ${event.operator}");
+          print(
+              "通过添加监听，获取到 loginAuthSyncApi 接口返回数据，code=${event.code},message = ${event.message},operator = ${event.operator}");
         });
+
         /// 再，执行同步的一键登录接口
         jverify.loginAuthSyncApi(autoDismiss: true);
       } else {
         setState(() {
-          _loading = false;
+          _hideLoading();
           _result = "[2016],msg = 当前网络环境不支持认证";
         });
 
@@ -445,7 +464,6 @@ class _MyAppState extends State<MyApp> {
         widgetList.add(buttonWidget);
         */
 
-
         /* 弹框模式
         JVPopViewConfig popViewConfig = JVPopViewConfig();
         popViewConfig.width = (screenWidth - 100.0).toInt();
@@ -453,7 +471,6 @@ class _MyAppState extends State<MyApp> {
 
         uiConfig.popViewConfig = popViewConfig;
         */
-
 
         /*
 
@@ -467,7 +484,7 @@ class _MyAppState extends State<MyApp> {
           String content = map[f_msg_key];
           String operator = map[f_opr_key];
           setState(() {
-            _loading = false;
+           _hideLoading();
             _result = "接口异步返回数据：[$code] message = $content";
           });
           print("通过接口异步返回，获取到 loginAuth 接口返回数据，code=$code,message = $content,operator = $operator");
@@ -479,30 +496,21 @@ class _MyAppState extends State<MyApp> {
     });
   }
 
-
-
-
   // Platform messages are asynchronous, so we initialize in an async method.
   Future<void> initPlatformState() async {
-    String platformVersion;
-
     // 初始化 SDK 之前添加监听
-    jverify.addSDKSetupCallBackListener((JVSDKSetupEvent event){
+    jverify.addSDKSetupCallBackListener((JVSDKSetupEvent event) {
       print("receive sdk setup call back event :${event.toMap()}");
     });
 
     jverify.setDebugMode(true); // 打开调试模式
     jverify.setup(
-        appKey: "a0e6ace8d5b3e0247e3f58db",//"你自己应用的 AppKey",
+        appKey: "a0e6ace8d5b3e0247e3f58db", //"你自己应用的 AppKey",
         channel: "devloper-default"); // 初始化sdk,  appKey 和 channel 只对ios设置有效
     // If the widget was removed from the tree while the asynchronous platform
     // message was in flight, we want to discard the reply rather than calling
     // setState to update our non-existent appearance.
     if (!mounted) return;
-
-    setState(() {
-      _platformVersion = platformVersion;
-    });
 
     /// 授权页面点击时间监听
     jverify.addAuthPageEventListener((JVAuthPageEvent event) {
@@ -513,25 +521,22 @@ class _MyAppState extends State<MyApp> {
 
 /// 封装 按钮
 class CustomButton extends StatelessWidget {
-
-  final VoidCallback onPressed;
-  final String title;
+  final VoidCallback? onPressed;
+  final String? title;
 
   const CustomButton({@required this.onPressed, this.title});
 
   @override
   Widget build(BuildContext context) {
-    // TODO: implement build
-    return new FlatButton(
+    return new TextButton(
       onPressed: onPressed,
       child: new Text("$title"),
-      color: Color(0xff585858),
-      highlightColor: Color(0xff888888),
-      splashColor: Color(0xff888888),
-      textColor: Colors.white,
-      padding: EdgeInsets.fromLTRB(10, 5, 10, 5),
+      style: ButtonStyle(
+        foregroundColor: MaterialStateProperty.all(Colors.white),
+        overlayColor: MaterialStateProperty.all(Color(0xff888888)),
+        backgroundColor: MaterialStateProperty.all(Color(0xff585858)),
+        padding: MaterialStateProperty.all(EdgeInsets.fromLTRB(10, 5, 10, 5)),
+      ),
     );
   }
 }
-
-
